Split setup steps by responsibility
