+++ Tasks/UnitySetup/UnitySetupV2/unity-editor-activation.ts
@@ -1,0 +1,4 @@
+export class UnityEditorActivation {
+    public static run(): number {
+    }
+}
